fixed logical bug on 2 specs in queryTests.js
Test fails in queryTests.js on both specs:
"on orders ordered after April '98 OR with freight > 100"
"orders ordered after April '98 AND with freight > 100"
The they're failing is because in JavaScript Date.toISOString() method
is somewhat ambiguous:
new Date(1998, 3, 1).toISOString() returns "1998-03-31T22:00:00.000Z"
This means that orders with OrderDate = "1998-04-01" are also included
with the queries.

To test this claim run this SQL on the server:
SELECT * FROM [Northwind].[dbo].[Order] where Freight > 100 and
OrderDate > '1998-03-31T22:00:00.000Z'
This query should return 16 instead of 15 records.